# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API de gestion commerciale (factures, bons de livraison, devis, articles, clients)
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type DocumentType = (typeof DocumentType)[keyof typeof DocumentType];

export const DocumentType = {
  facture: "facture",
  devis: "devis",
  bon_livraison: "bon_livraison",
} as const;

export type DocumentStatus =
  (typeof DocumentStatus)[keyof typeof DocumentStatus];

export const DocumentStatus = {
  brouillon: "brouillon",
  valide: "valide",
  livre: "livre",
  paye: "paye",
  annule: "annule",
} as const;

export interface Company {
  id: number;
  name: string;
  address: string;
  phone: string;
  fiscalNumber: string;
  rccm: string;
  /** Comptes bancaires (texte libre) */
  bankAccounts: string;
  comptoirName: string;
  comptoirCity: string;
  comptoirPhone: string;
  tvaRate: number;
  currency: string;
  legalFooter: string;
}

export interface UpdateCompanyBody {
  name: string;
  address: string;
  phone: string;
  fiscalNumber: string;
  rccm: string;
  bankAccounts: string;
  comptoirName: string;
  comptoirCity: string;
  comptoirPhone: string;
  tvaRate: number;
  currency: string;
  legalFooter: string;
}

export interface Client {
  id: number;
  name: string;
  contactName?: string | null;
  address?: string | null;
  city?: string | null;
  phone?: string | null;
  email?: string | null;
  fiscalNumber?: string | null;
  notes?: string | null;
  createdAt: string;
}

export interface CreateClientBody {
  name: string;
  contactName?: string | null;
  address?: string | null;
  city?: string | null;
  phone?: string | null;
  email?: string | null;
  fiscalNumber?: string | null;
  notes?: string | null;
}

export interface Article {
  id: number;
  reference: string;
  designation: string;
  unite: string;
  prixUnitaire: number;
  depot?: string | null;
  stock: number;
  createdAt: string;
}

export interface CreateArticleBody {
  reference: string;
  designation: string;
  unite: string;
  prixUnitaire: number;
  depot?: string | null;
  stock: number;
}

export interface DocumentLine {
  id: number;
  articleId?: number | null;
  reference: string;
  designation: string;
  quantite: number;
  unite: string;
  prixUnitaire: number;
  remisePct: number;
  montantHt: number;
  depot?: string | null;
}

export interface CreateDocumentLineBody {
  articleId?: number | null;
  reference: string;
  designation: string;
  quantite: number;
  unite: string;
  prixUnitaire: number;
  remisePct: number;
  depot?: string | null;
}

export interface Document {
  id: number;
  type: DocumentType;
  numero: string;
  status: DocumentStatus;
  date: string;
  echeance?: string | null;
  clientId: number;
  client: Client;
  vendeur?: string | null;
  reference?: string | null;
  notes?: string | null;
  lines: DocumentLine[];
  totalHt: number;
  totalRemise: number;
  totalTva: number;
  totalTtc: number;
  applyTva: boolean;
  relatedDocumentId?: number | null;
  createdAt: string;
}

export interface DocumentSummary {
  id: number;
  type: DocumentType;
  numero: string;
  status: DocumentStatus;
  date: string;
  echeance?: string | null;
  clientId: number;
  clientName: string;
  totalTtc: number;
  createdAt: string;
}

export interface CreateDocumentBody {
  type: DocumentType;
  date: string;
  echeance?: string | null;
  clientId: number;
  vendeur?: string | null;
  reference?: string | null;
  notes?: string | null;
  applyTva: boolean;
  status: DocumentStatus;
  lines: CreateDocumentLineBody[];
}

export interface UpdateDocumentStatusBody {
  status: DocumentStatus;
}

export interface ConvertDocumentBody {
  targetType: DocumentType;
}

export interface DashboardSummary {
  totalRevenue: number;
  totalUnpaid: number;
  invoiceCount: number;
  quoteCount: number;
  deliveryCount: number;
  clientCount: number;
  articleCount: number;
  currency: string;
}

export interface TopClient {
  clientId: number;
  clientName: string;
  totalRevenue: number;
  invoiceCount: number;
}

export interface TopArticle {
  articleId?: number | null;
  reference: string;
  designation: string;
  totalQuantity: number;
  totalRevenue: number;
}

export interface MonthlyRevenue {
  /** YYYY-MM */
  month: string;
  revenue: number;
  invoiceCount: number;
}

export type ListClientsParams = {
  search?: string;
};

export type ListArticlesParams = {
  search?: string;
};

export type ListDocumentsParams = {
  type?: DocumentType;
  status?: DocumentStatus;
  clientId?: number;
  search?: string;
};

export type GetRecentDocumentsParams = {
  limit?: number;
};

export type GetTopClientsParams = {
  limit?: number;
};

export type GetTopArticlesParams = {
  limit?: number;
};