# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API de gestion commerciale (factures, bons de livraison, devis, articles, clients)
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /clients:
    get:
      operationId: listClients
      tags: [clients]
      summary: Lister tous les clients
      parameters:
        - name: search
          in: query
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Client"
    post:
      operationId: createClient
      tags: [clients]
      summary: Créer un client
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateClientBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Client"
  /clients/{id}:
    get:
      operationId: getClient
      tags: [clients]
      parameters:
        - $ref: "#/components/parameters/IdParam"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Client"
    put:
      operationId: updateClient
      tags: [clients]
      parameters:
        - $ref: "#/components/parameters/IdParam"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateClientBody"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Client"
    delete:
      operationId: deleteClient
      tags: [clients]
      parameters:
        - $ref: "#/components/parameters/IdParam"
      responses:
        "204":
          description: Deleted
  /articles:
    get:
      operationId: listArticles
      tags: [articles]
      summary: Lister tous les articles
      parameters:
        - name: search
          in: query
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Article"
    post:
      operationId: createArticle
      tags: [articles]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateArticleBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Article"
  /articles/{id}:
    get:
      operationId: getArticle
      tags: [articles]
      parameters:
        - $ref: "#/components/parameters/IdParam"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Article"
    put:
      operationId: updateArticle
      tags: [articles]
      parameters:
        - $ref: "#/components/parameters/IdParam"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateArticleBody"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Article"
    delete:
      operationId: deleteArticle
      tags: [articles]
      parameters:
        - $ref: "#/components/parameters/IdParam"
      responses:
        "204":
          description: Deleted
...
  /dashboard/recent-documents:
    get:
      operationId: getRecentDocuments
      tags: [dashboard]
      summary: Documents récents
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 10
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/DocumentSummary"
  /dashboard/top-clients:
    get:
      operationId: getTopClients
      tags: [dashboard]
      summary: Clients les plus actifs (par chiffre d'affaires)
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 5
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TopClient"
  /dashboard/top-articles:
    get:
      operationId: getTopArticles
      tags: [dashboard]
      summary: Articles les plus vendus
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 5
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TopArticle"
...
components:
  parameters:
    IdParam:
      name: id
      in: path
      required: true
      schema:
        type: integer
  schemas:
...
    DocumentType:
      type: string
      enum: [facture, devis, bon_livraison]
    DocumentStatus:
      type: string
      enum: [brouillon, valide, livre, paye, annule]
...
    Client:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        contactName:
          type: string
          nullable: true
        address:
          type: string
          nullable: true
        city:
          type: string
          nullable: true
        phone:
          type: string
          nullable: true
        email:
          type: string
          nullable: true
        fiscalNumber:
          type: string
          nullable: true
        notes:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
      required: [id, name, createdAt]
    CreateClientBody:
      type: object
      properties:
        name:
          type: string
        contactName:
          type: string
          nullable: true
        address:
          type: string
          nullable: true
        city:
          type: string
          nullable: true
        phone:
          type: string
          nullable: true
        email:
          type: string
          nullable: true
        fiscalNumber:
          type: string
          nullable: true
        notes:
          type: string
          nullable: true
      required: [name]
    Article:
      type: object
      properties:
        id:
          type: integer
        reference:
          type: string
        designation:
          type: string
        unite:
          type: string
        prixUnitaire:
          type: number
          format: float
        depot:
          type: string
          nullable: true
        stock:
          type: number
          format: float
        createdAt:
          type: string
          format: date-time
      required: [id, reference, designation, unite, prixUnitaire, stock, createdAt]
    CreateArticleBody:
      type: object
      properties:
        reference:
          type: string
        designation:
          type: string
        unite:
          type: string
        prixUnitaire:
          type: number
          format: float
        depot:
          type: string
          nullable: true
        stock:
          type: number
          format: float
      required: [reference, designation, unite, prixUnitaire, stock]
...
    DocumentSummary:
      type: object
      properties:
        id:
          type: integer
        type:
          $ref: "#/components/schemas/DocumentType"
        numero:
          type: string
        status:
          $ref: "#/components/schemas/DocumentStatus"
        date:
          type: string
          format: date
        echeance:
          type: string
          format: date
          nullable: true
        clientId:
          type: integer
        clientName:
          type: string
        totalTtc:
          type: number
          format: float
        createdAt:
          type: string
          format: date-time
      required: [id, type, numero, status, date, clientId, clientName, totalTtc, createdAt]
...
    TopClient:
      type: object
      properties:
        clientId:
          type: integer
        clientName:
          type: string
        totalRevenue:
          type: number
          format: float
        invoiceCount:
          type: integer
      required: [clientId, clientName, totalRevenue, invoiceCount]
    TopArticle:
      type: object
      properties:
        articleId:
          type: integer
          nullable: true
        reference:
          type: string
        designation:
          type: string
        totalQuantity:
          type: number
          format: float
        totalRevenue:
          type: number
          format: float
      required: [reference, designation, totalQuantity, totalRevenue]
...