# lib/api-zod/src/generated/types/createArticleBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API de gestion commerciale (factures, bons de livraison, devis, articles, clients)
 * OpenAPI spec version: 0.1.0
 */

export interface CreateArticleBody {
  reference: string;
  designation: string;
  unite: string;
  prixUnitaire: number;
  depot?: string | null;
  stock: number;
}