# lib/api-zod/src/generated/types/topClient.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API de gestion commerciale (factures, bons de livraison, devis, articles, clients)
 * OpenAPI spec version: 0.1.0
 */

export interface TopClient {
  clientId: number;
  clientName: string;
  totalRevenue: number;
  invoiceCount: number;
}